requirements.txt
dj-database-url==0.3.0
Django==1.7.4
django-configurations==0.8
django-debug-toolbar==1.2.2
django-dotenv==1.3.0
django-extensions==1.5.0
django-secure==1.0.1
gunicorn==19.2.1
# newrelic==2.42.0.35
# psycopg2==2.5.4
Werkzeug==0.10.1
django-imagekit
django-contact-form
svgwrite
slumber
django-braces==1.4.0
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
django-extra-views==0.6.5
django-filter==0.9.1
django-floppyforms==1.2.0
djangorestframework==3.0
pytz==2014.7
pillow==2.7.0